+++ lib/esse/events.rb
@@ -58,5 +58,6 @@ module Esse
     register_event 'elasticsearch.get'
     register_event 'elasticsearch.reindex'
     register_event 'elasticsearch.update_by_query'
+    register_event 'elasticsearch.delete_by_query'
   end
 end